Move consumers into their own directory
